wrapper/requirements.txt
Flask==3.0.0
google-cloud-pubsub==2.19.0
requests==2.31.0
gunicorn==21.2.0
